Conflicts:
	src/mongo/client/examples/mongoperf.vcxproj
	src/mongo/client/examples/mongoperf.vcxproj.filters
	src/mongo/db/curop.h
	src/mongo/db/db.vcxproj.filters
	src/mongo/dbtests/queryoptimizercursortests.cpp
	src/mongo/dbtests/test.vcxproj.filters
	src/mongo/dbtests/threadedtests.cpp
	src/mongo/shell/msvc/mongo.vcxproj
	src/mongo/shell/msvc/mongo.vcxproj.filters

+++ src/mongo/db/pipeline/expression.cpp
@@ -1831,7 +1831,6 @@ namespace mongo {
         const intrusive_ptr<const Value> &pValue):
         pFieldPath(pTheFieldPath),
         pRange(new Range(cmpOp, pValue)) {
-
     }
 
     void ExpressionFieldRange::intersect(
@@ -2066,7 +2065,7 @@ namespace mongo {
         ExpressionNary::addOperand(pExpression);
     }
 
-    int ExpressionIsoDate::checkIntRange(const char *pName, long value) const {
+    int ExpressionIsoDate::checkIntRange(const char *pName, long long value) const {
         uassert(16027, str::stream() << getOpName() <<
                 ":  \"" << pName <<
                 "\" value outside of range INT_MIN to INT_MAX",
